requirements.txt
pytest==6.2.4
selenium==3.141.0
webdriver-manager==3.4.2
allure-pytest==2.9.43
allure-python-commons==2.9.43
